+++ src/laser/init/transformers/unwpp.py
@@ -1,0 +1,15 @@
+"""
+Transform UNWPP demographic data to a usable format, filtered by country and year range.
+"""
+
+
+class UnwppTransformer:
+    def __init__(self):
+        pass
+
+    @staticmethod
+    def description():
+        return "Transform UNWPP demographic data to a usable format, filtered by country and year range."
+
+    def transform(self, stats_data, iso_code, start_year, end_year):
+        return